Float: Always use 32 bits integers for integer operations.
I think it is important that 'compl 0' gives the same result
whatever the platform is.

On a side note, this syntax is not consistant with the rest of
the language.

	* doc/specs/float.tex: Mention the fact that compl
	uses 32 bits integer.
	* src/object/cxx-conversions.hxx: Adapt.
	* src/object/float.hh: switch unsigned_type to unsigned int.

+++ src/object/cxx-conversions.hxx
@@ -154,7 +154,7 @@ namespace object
 
   CONVERT(unsigned char);
   CONVERT(unsigned short);
-  CONVERT(unsigned int);
+  CONVERT(unsigned long);
 #undef CONVERT
 
   /*--------.
